Problem: YLR doesn't specify entity serialization
Solution: specify entity serialization based on layout's structure

Also, rehaul layout serialization to indicate the causation between
fingerprint and layout name (layout name is part of the fingerprint,
not the other way around). The necessity of this change was highlighted
by the need for the entity to refer to layout's fingerprint, not
its name.